Using longlong instead of long long. :-D
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21211 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/aws/awscomp.cpp
@@ -391,7 +391,7 @@ bool awsComponent::Execute (const char* action, iAwsParmList* parmlist)
 
 	if (f.parseObject(code.begin(), code.end()))
 	{
-		f.addIntParm("comp_id", (long long)self);
+		f.addIntParm("comp_id", (longlong)self);
 		autom::keeper result = f.Execute();
 	}
